+++ docs/source/release-notes.rst
@@ -13,6 +13,7 @@ v1.1.1 (2025-xx-xx)
 ----------------------
 
 - Include objective function value in exported report
+- Add PTDF-based routines: ``RTED2``, ``RTEDDG2``, ``RTEDES2``
 
 v1.1.0 (2025-12-18)
 ----------------------